requirements.txt
black==24.3.0
flake8==5.0.4
mock==4.0.3
pre-commit==2.20.0
pytest==7.1.3
apache-airflow==2.10.1
apache-airflow-providers-airbyte==4.0.0
apache-airflow-providers-github== 2.7.0
pandas==2.2.3
